requirements.txt
alembic==1.0.11
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.7.0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
dominate==2.4.0
entrypoints==0.3
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.720
mypy-extensions==0.4.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
alembic==1.0.11
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.7.0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
dominate==2.4.0
elasticsearch==7.0.4
entrypoints==0.3
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.720
mypy-extensions==0.4.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
